Moving get_convex_on_image to utils.cpp
