+++ Advancement.md
@@ -12,9 +12,9 @@
 - [x] class `Ray` &rarr; `Ray.hpp/.cpp`
 - [x] class `Ray_Hash` &rarr; `Ray_Hash.hpp/.cpp`
 - [x] class `Ray_Information` &rarr; `Ray_Information.hpp/.cpp`
-- [ ] `get_xmax_xmin_ymax_ymin_in_hull` &rarr; `utils.cpp`
-- [ ] `is_pixel_in_convex` &rarr; `utils.cpp`
-- [ ] `get_convex_on_image` &rarr; `utils.cpp`
+- [x] `get_xmax_xmin_ymax_ymin_in_hull` &rarr; `utils.cpp`
+- [x] `is_pixel_in_convex` &rarr; `utils.cpp`
+- [x] `get_convex_on_image` &rarr; `utils.cpp`
 - [ ] `project_pixel_to_ray_end` &rarr; `utils.cpp`
 - [ ] `frontier_check` &rarr; `utils.cpp`
 - [ ] `distance_function` &rarr; `utils.cpp`
